requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
openai>=1.0
python-dotenv>=1.0
pyyaml>=6.0
sqlmodel>=0.0.16
sqlalchemy>=2.0
python-multipart>=0.0.9
